refactor: extract commands module structure. Closes #12
Move validate command logic into src/commands/validate.rs and
src/commands/mod.rs. main.rs is now thin: only CLI arg parsing
and dispatch.
